blockchain/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.5
crypto==1.4.1
Django==3.2.13
idna==3.4
Naked==0.1.32
pycryptodome==3.16.0
PyMySQL==1.0.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
sqlparse==0.4.3
urllib3==1.26.13
